+++ CHANGELOG.rst
@@ -38,6 +38,7 @@ in development
 * Turn on paramiko ssh runner as the default ssh runner in prod configuration.
   To switch to fabric runner, set ``use_paramiko_ssh_runner`` to false in st2.conf. (improvement)
 * Add OpenStack Keystone authentication configuration for Mistral. (improvement)
+* Add ability to add trace tag to ``st2 run`` CLI command. (feature)
 
 0.12.2 - August 11, 2015.
 -------------------------